Fix mkdir and state issues

+++ components/bridge/synchronization/impl/src/main/java/com/flipperdevices/bridge/synchronization/impl/executor/DiffKeyExecutor.kt
@@ -47,7 +47,7 @@ class DiffKeyExecutorImpl @Inject constructor() : DiffKeyExecutor, LogTagProvide
             try {
                 info { "Execute $diff for $source to $target" }
                 execute(source, target, diff)
-                tracker.report(index, diffs.size)
+                tracker.report(index.toLong(), diffs.size.toLong())
                 return@mapIndexedNotNull diff
             } catch (executeError: Exception) {
                 error(executeError) { "While apply diff $diff we have error" }
